requirements.txt
pytest
pytest-cov
voluptuous
pyyaml
-e .
